Revert "Bug: 22290"
This reverts commit ec972df029126c89dee4bfeca4b02a1990311cc9.

+++ src/SME.SGP.Dados/Repositorios/RepositorioFrequenciaAlunoDisciplinaPeriodo.cs
@@ -22,7 +22,7 @@ namespace SME.SGP.Dados.Repositorios
                         from
 	                        frequencia_aluno
                         where
-	                        codigo_aluno = @codigoAluno
+	                        codigo_aluno = @codigoAluno::int8
 	                        and disciplina_id = @disciplinaId
 	                        and tipo = @tipoFrequencia
 	                        and periodo_escolar_id = @periodoEscolarId";
@@ -100,7 +100,7 @@ namespace SME.SGP.Dados.Repositorios
             var query = new StringBuilder(@"select * 
                             from frequencia_aluno
                            where tipo = 2
-	                        and codigo_aluno = @alunoCodigo
+	                        and codigo_aluno = @alunoCodigo::int8
                             and turma_id = @turmaCodigo ");
 
             if (!string.IsNullOrEmpty(componenteCurricularCodigo))
@@ -118,7 +118,7 @@ namespace SME.SGP.Dados.Repositorios
         public async Task<IEnumerable<FrequenciaAluno>> ObterFrequenciaBimestresAsync(string codigoAluno, int bimestre, string codigoTurma)
         {
             var query = @"select * from frequencia_aluno fa 
-                            where fa.codigo_aluno = @codigoAluno
+                            where fa.codigo_aluno = @codigoAluno::int8
                             and fa.turma_id = @turmaId and fa.tipo = 1";
 
             if (bimestre > 0)
@@ -138,7 +138,7 @@ namespace SME.SGP.Dados.Repositorios
         {
             var query = new StringBuilder(@"select *
                         from frequencia_aluno
-                        where codigo_aluno = @codigoAluno
+                        where codigo_aluno = @codigoAluno::int8
 	                        and tipo = @tipoFrequencia
 	                        and bimestre = @bimestre ");
 
@@ -161,7 +161,7 @@ namespace SME.SGP.Dados.Repositorios
                         from frequencia_aluno fa
                         inner join periodo_escolar pe on fa.periodo_escolar_id = pe.id
                         where
-	                        codigo_aluno = @codigoAluno
+	                        codigo_aluno = @codigoAluno::int8
 	                        and tipo = @tipoFrequencia
 	                        and pe.periodo_inicio <= @dataAtual
 	                        and pe.periodo_fim >= @dataAtual ");
@@ -183,7 +183,7 @@ namespace SME.SGP.Dados.Repositorios
             var query = @"select *
                         from frequencia_aluno fa
                         inner join periodo_escolar pe on fa.periodo_escolar_id = pe.id
-                        where codigo_aluno = @codigoAluno
+                        where codigo_aluno = @codigoAluno::int8
                             and disciplina_id = @disciplinaId
 	                        and tipo = 1
 	                        and pe.periodo_inicio <= @dataAtual
